reorganizing the paging module
