+++ core/repository/paging/src/main/java/social/firefly/core/repository/paging/remotemediators/HomeTimelineRemoteMediator.kt
@@ -1,4 +1,4 @@
-package social.firefly.core.repository.paging
+package social.firefly.core.repository.paging.remotemediators
 
 import androidx.paging.ExperimentalPagingApi
 import androidx.paging.LoadType
