+++ feature/discover/src/main/java/social/firefly/feature/discover/DiscoverViewModel.kt
@@ -3,13 +3,10 @@ package social.firefly.feature.discover
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import androidx.paging.ExperimentalPagingApi
-import androidx.paging.PagingData
 import androidx.paging.map
-import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.launch
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
 import org.koin.core.parameter.parametersOf
@@ -19,17 +16,13 @@ import social.firefly.core.analytics.FeedLocation
 import social.firefly.core.navigation.NavigationDestination
 import social.firefly.core.navigation.usecases.NavigateTo
 import social.firefly.core.repository.mastodon.TrendingHashtagRepository
-import social.firefly.core.repository.paging.TrendingHashtagsRemoteMediator
-import social.firefly.core.repository.paging.TrendingStatusPager
-import social.firefly.core.ui.common.following.FollowStatus
+import social.firefly.core.repository.paging.remotemediators.TrendingHashtagsRemoteMediator
+import social.firefly.core.repository.paging.pagers.TrendingStatusPager
 import social.firefly.core.ui.hashtagcard.HashTagCardDelegate
 import social.firefly.core.ui.hashtagcard.quickview.toHashTagQuickViewUiState
 import social.firefly.core.ui.postcard.PostCardDelegate
 import social.firefly.core.ui.postcard.toPostCardUiState
 import social.firefly.core.usecase.mastodon.account.GetLoggedInUserAccountId
-import social.firefly.core.usecase.mastodon.hashtag.FollowHashTag
-import social.firefly.core.usecase.mastodon.hashtag.UnfollowHashTag
-import timber.log.Timber
 
 @OptIn(ExperimentalPagingApi::class)
 class DiscoverViewModel(
