+++ feature/hashtag/src/main/java/social/firefly/feature/hashtag/HashTagViewModel.kt
@@ -19,7 +19,7 @@ import social.firefly.core.analytics.FeedLocation
 import social.firefly.core.analytics.HashtagAnalytics
 import social.firefly.core.model.HashTag
 import social.firefly.core.repository.mastodon.TimelineRepository
-import social.firefly.core.repository.paging.HashTagTimelineRemoteMediator
+import social.firefly.core.repository.paging.remotemediators.HashTagTimelineRemoteMediator
 import social.firefly.core.ui.postcard.PostCardDelegate
 import social.firefly.core.ui.postcard.toPostCardUiState
 import social.firefly.core.usecase.mastodon.account.GetLoggedInUserAccountId
