+++ feature/settings/src/main/java/social/firefly/feature/settings/contentpreferences/blockedusers/BlockedUsersViewModel.kt
@@ -14,7 +14,7 @@ import social.firefly.core.analytics.SettingsAnalytics
 import social.firefly.core.model.BlockedUser
 import social.firefly.core.navigation.usecases.NavigateToAccount
 import social.firefly.core.repository.mastodon.BlocksRepository
-import social.firefly.core.repository.paging.BlocksListRemoteMediator
+import social.firefly.core.repository.paging.remotemediators.BlocksListRemoteMediator
 import social.firefly.core.ui.common.account.quickview.toQuickViewUiState
 import social.firefly.core.ui.common.account.toggleablelist.ToggleableAccountListItemState
 import social.firefly.core.usecase.mastodon.account.BlockAccount
